+++ SoftwareManual/relative_error.cpp
@@ -2,18 +2,18 @@
 #include <cmath>
 using namespace std;
 
-int relativeError (int solution, int computedSolution);
+float relativeError (float solution, float computedSolution);
 
 int main() {
-    int solution;
-    int computedSolution;
+    float solution;
+    float computedSolution;
     cin >> solution;
     cin >> computedSolution;
     cout << relativeError(solution,computedSolution);
     return 0;
 }
 
-int relativeError (int solution, int computedSolution) {
-    int relativeError = abs((computedSolution - solution)/(solution));
+float relativeError (float solution, float computedSolution) {
+    float relativeError = abs((computedSolution - solution)/(solution));
     return relativeError;
 }
